requirements.txt
APScheduler==3.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
greenlet==1.1.2
idna==3.3
numpy==1.22.3
pandas==1.4.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
urllib3==1.26.9
